Update to support new ddc apis
refresh_candidates was deprecated and deleted, which relegated this
plugin to being completely broken. I've updated it to support the new
callback api.  I'd like to give it a day or two to test, just to ensure
nothing subtle is broken.
